pyproject.toml
[project]
name = "catgpt"
version = "0.1.0"
description = "ML research project for chess and beyond"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "CatGPT Team" }]
keywords = ["ml", "research", "chess", "deep-learning", "pytorch", "jax"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies (framework-agnostic)
dependencies = [
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "python-chess>=1.999",
    "tqdm>=4.66.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
]

[project.optional-dependencies]
# PyTorch framework
torch = [
    "torch>=2.2.0",
]
# JAX framework
jax = [
    "jax[cpu]>=0.4.20",
    "flax>=0.8.0",
    "optax>=0.2.0",
    "orbax-checkpoint>=0.5.0",
]
# JAX with CUDA support
jax-cuda = [
    "jax[cuda12]>=0.4.20",
    "flax>=0.8.0",
    "optax>=0.2.0",
    "orbax-checkpoint>=0.5.0",
]
# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "pyright>=1.1.350",
    "pre-commit>=3.6.0",
    "ipython>=8.0.0",
]
# Jupyter notebooks
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]
# All frameworks and tools
all = ["catgpt[torch,jax,dev,notebook]"]

[project.scripts]
catgpt = "catgpt.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/catgpt"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]
extend-exclude = [
    "src/catgpt/torch/optimizers/splus.py",  # Third-party file
    "src/catgpt/jax/*",  # JAX code excluded until JAX is installed
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["catgpt"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
include = ["src"]
exclude = [
    "**/__pycache__",
    "build",
    "dist",
    ".venv",
    "src/catgpt/torch/optimizers/splus.py",
    "src/catgpt/jax",  # JAX code excluded until JAX is installed
]
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "jax: marks tests that require JAX",
]

[tool.coverage.run]
source = ["src/catgpt"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*", "src/catgpt/jax/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
